requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
cryptography==43.0.1
cymem==2.0.8
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
faiss-cpu==1.9.0
fastapi==0.115.2
filelock==3.16.1
fsspec==2024.9.0
h11==0.14.0
httptools==0.6.4
huggingface-hub==0.25.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.1
language_data==1.2.0
loguru==0.7.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.4.1
nltk==3.9.1
numpy==2.0.2
packaging==24.1
pdfminer.six==20240706
pillow==11.0.0
preshed==3.0.9
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.0
shellingham==1.5.4
smart-open==7.0.5
sniffio==1.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
starlette==0.40.0
sympy==1.13.3
thinc==8.3.2
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.4.1
tqdm==4.66.5
transformers==4.45.2
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
wasabi==1.1.3
watchfiles==0.24.0
weasel==0.4.1
websockets==13.1
wrapt==1.16.0
